Add a sample macOS embedder application
Initial bare-bones macOS example app.

Currently this shows only a simple hello-world application; it will be
expanded in the future to demonstrate other capabilities like text
handling, eventually menu support, etc.

It also currently makes no attempt to package the necessary Flutter code
into the application, so can't be moved/distributed. Packaging will be
added in the future, but this provides a simple starting point for
getting code up and running.

By design, this has minimal changes at this point from the standard
project created by Xcode, so many build settings are not important to
the example. In the future, project and target settings may be pared
down to make it more clear what's specific to making a working embedder.

Known issue: Font resolution does not appear to be wired up correctly
(even if the MaterialIcons font is loaded), so the button in the sample
app does not display the '+'.

Closes #3