Add support for snapshot mode to FLEViewController
The embedding API allows passing empty main and package paths to use
snapshot mode rather than source mode, which allows for easy bundling.

Switches the example app to use snapshot mode, and bundles the build
resources directory, so that the app is self-contained. Removes the
need for the command-line argument workaround (which was accidentally
missing from the previous patch due to the scheme that passed the path
to the example app directory not being marked as shared, so this has
the added effect of fixing the sample app).

Closes #8