Convert the Flutter engine fetcher to Dart
Replaces the bash script that fetches a prebuilt Flutter engine with a
Dart version, to make it more portable and easier to maintain.

Adds support for downloading Linux and Windows libraries in addition to
macOS. (The script has not yet been tested on those platforms, but
all three target options work on macOS.)

Bundles all dependencies of the script in the repository to avoid
forcing a pub step during setup.