Use update_flutter_engine in the Linux build
Instead of requiring that the Linux engine be manually added to the tree,
fetch the correct prebuilt version automatically as is done on macOS.

Updates the source to expect the engine's header to be called
flutter_embedder.h, as that is what it is named in the archive. The
desktop embedding header has been renamed from flutter_embedder.h
to just embedder.h to avoid confusion.

Fixes issue #62