Make the color panel a standalone plugin
This extracts the color panel out of the embedder frameworks and into
a standalone plugin. As part of that, adds the Dart side of the plugin
to the respository.

Since this is the first plugin extracted, it serves as a template for
the plugin structure:
- Creates a new top-level plugins/ directory, with README.
- The plugin is a subdirectory containing:
  - The Dart code as a (local) package.
  - The macOS and Linux code, in macos/ and linux/ subdirectories
    (this code is moved from the framework directories).
    - Each has build files to create the native plugin library.
  - A common/ subdirectory for channel constants shared across
    platforms.
    - The duplicate strings have been extracted from the native
      code into a shared file.

Since the Dart code is now part of the repository, this patch also
does some queued cleanup that required two-sided changes:
- Removes the unnecessary wrapping array from the color callback.
- Moves the normalization to 0-255 to the Dart side, so that it's
  not duplicated in the native code for each platform.
- Adds a callback for the case when the user closes the panel, so
  that the Dart-side understanding of whether the panel is showing
  stays accurate.

Fixes issues #9