Add a configuration option for Flutter location
Requiring the Flutter tree to be in a specific location is problematic
for client projects that need to control layouts (e.g., need to map
Flutter at another location for other reasons), and also makes it
difficult for individuals to share the tree used for this project with
other uses of Flutter.

To improve flexibility, allow a .flutter_location_config file in the
directory containing this repository (not inside it, so that it can be
checked in to an enclosing repository if desired) that specifies the
location of the Flutter tree.