Set an explicit platform and font for the example
This forces a target platform instead of relying on a host->target
mapping internal to Flutter (Fuchsia is used since it's not the
default for any desktop host platform, so puts everything on an
even footing), as well as setting the font to Roboto and bundling
Roboto.

This will make the example more consistent across platforms, and
should address Issue #57.