Set debugDefaultTargetPlatformOverride
Flutter no longer maps desktop host platforms to mobile target
platforms, and instead throws 'Unknown platform' exceptions.
This adds a debugDefaultTargetPlatformOverride to bypass the
detection logic so that the example runs on desktop.

Also updates the main README to include a section on likely pitfalls
when attempting to run a stock Flutter application on desktop (so far,
this issue and font handling), along with suggested solutions.