[macOS] Extract codecs from plugin API
Extracts method codecs into stand-alone objects, and allows plugins to
indicate that they use a codec other than JSON. This is an incremental
step toward aligning the plugin API with Flutter's MethodChannel API.

For now no other codecs are actually implemented in the framework, and
the default is still JSON (unlike in existing Flutter code) so that this
is not a breaking change. A future change will full move to MethodChannel,
which will be a breaking change.

Partially addresses issue #67 in that custom codecs are now supported,
even though the Standard (binary) codec is not yet implemented.