[linux] Split channels.h into separate files
Splits MethodCall and MethodResult into separate files. This follows the
more usual C++ convention of having one file per class, and makes the
code structure more consistent with the Android Flutter plugin
implementation (vs. following the iOS file structure, as it did before).

This is purely a move of code, with no implementation changes.
It is a precursor to the next plugin refactor, to avoid having
substantial functionality change mixed with file moves.

JsonMethodResult is left with MethodResult since it will be eliminated
in the refactor anyway.