[macOS] Align FLEPlugin with iOS Flutter
Changes FLEPlugin to be a subset of iOS's FlutterPlugin's API, including
changing the plugin creation and registration system. Plugins now create
and manage their own channels, instead of relying on FLEViewController
for high-level message APIs.

This is a breaking change for any FLEPlugin implementation, as well as
application-level code that was calling addPlugin:. (The later may
require additional changes in the future as noted in the comment on
registerPlugin:, but the plugin implemention API should now be stable.)

This essentially completes issue #102 for macOS.