Update pubspecs
Add compatibility up to Dart 3.0, to fix building with Flutter head,
which now uses 2.1 (with nothing specified, <2 is assumed as the
requirement).

Also removes the example's pubspeck.lock and adds it to the ignore
list until the project has a plan in place for specifying the required
version of Flutter.