Make the Linux example a moveable bundle
Instead of using libraries and resources from their source tree
locations, bundles them all into a directory with the binary so that it
can be run independently of the source.

Also changes the lookup of resources to be relative to the binary
(using /proc/self/exe) so that it's not sensitive to the working
direcotry it's run with. This is not fully portable, but alternative
lookup methods can be added in the future behind platform ifdefs.