[macos] Add MethodChannel
Adds the MethodChannel class, and the accompanying BinaryMessegener
protocol, for parity with the existing iOS Flutter plugin/channel
structure.

This lays the foundation for a future breaking change that makes plugin
creation and registration match the iOS structure, but for now makes the
method channel an internal implementation detail of the existing plugin
messaging design to preserve compatibiility.