[macos] Move FLEMethodCall and FLEMethodError (#128)
Moves FLEMethodCall and FLEMethodError out of Channels.h. This better
aligns with the Flutter iOS structure, although this deviates in that on
iOS these classes live in the Codecs.* files, while this keeps them
separate to reduce the number of classes being declared in a single
header.

Preparation for the remaining plugin refactoring (which adds channel
classes to the Channels.* files).