[linux] Add the beginning of BinaryMessenger
Adds the BinaryMessanger interface, containing only the send API for
now. This eliminates the engine pointer from a public header, and is a
small step toward moving toward supporting MessageChannel/MethodChannel.

Technically this is a breaking change, but there's no expectation that
any code outside of the library would be using set_engine, so should not
actually impact anyone.

Part of issue #102.