[linux] Add MethodChannel
This adds the MethodChannel and related APIs on Linux, further aligning
with the Flutter APIs. As with the current state on macOS, this is
implemented as an implementation detail of the existing plugin classes,
so that there's no plugin API breakage at this stage.

This currently deviates from the MethodChannel API on other platforms by
using InvokeMethodCall instead of InvokeMethod. Addressing this will be
done in a follow-up change.

Part of issue #102