Move most C++ library code to common/ and include/
This is a preperatory change for Windows using all of this code. Since
it will be shared between multiple platforms, it should live in a shared
location.

Because the GN build flattens the headers in the generated out/include/,
some files have #if/#else branches for #includes. Once there is a final
decision on whether to use GN, this can be re-evaluated.