[linux/windows] Align C++ plugin structure with Flutter (#183)
Eliminates Plugin as the primary interface point, and instead moves to a
MethodChannel-based model. This eliminates the need for void* types in
the interfaces, allowing a move to full type safety (via templates) for
all the classes related to channels.

Additionally, the plugin creation and registration system now matches
the Flutter structure on iOS and Android, using static registration of a
plugin class with a PluginRegistrar. The desktop registrar currently has
minimal functionality, but it can be extended over time.

This is a breaking change for any Plugin implementation, as well as
application-level code that was calling AddPlugin.

This essentially completes issue #102 for Linux and Windows.