Update the source for fetch_jsoncpp
Changes fetch_jsoncpp to use the main jsoncpp repository instead of the
clarkezone fork, now that the necessary VS support has landed in the
main repo.

Fixes current Windows build errors due to the previously-pinned commit
no longer existing due to a force-push in the forked repository.

Adds temporary logic to fetch_jsoncpp to try to fix up existing
checkouts. (Also removes the assertions that fetch_jsoncpp is being run
on Windows since it's not Windows-specific, and it's easier to maintain
if it can be run on any platform.)