Re-organize the example app
Instead of making the platform implementations peers of the example
Flutter application, make them subdirectories of it. This matches the
structure of a standard Flutter application's mobile implementations,
and provides a better example of how to use the project to add desktop
support to an existing Flutter app.

Since interest so far has been for that use case rather than adding
Flutter to an existing application (and the current Linux and Windows
implementations don't really support the hybrid use case anyway), this
should be more reflective of how it would actually be used.