Add Flutter tree min version check
To minimize confusion, and reduce unnecessary issue reports, add a check
during builds that the Flutter tree being used is not older than the
last known point where there was a compatibility break. Instead of a
runtime error, or an unexplained build error, this will fail out during
the engine update step with a clear message to use a newer version of
Flutter.

Currently, this will ensure that people who aren't on Flutter master
will get an error message that's as clear as possible.