[linux] Switch to GN, and remove Makefiles
For the library and plugins, switch to GN as the only supported build
system on Linux. GN will be used for the library in the flutter/engine
repository, so the make build will no longer be necessary; remove it now
to avoid the overhead of maintaining an extra build system until the
transition happens.

The example continues to use Make in order to show typical usage of the
library, and the color_panel plugin has an example makefile (updated to
reflect the use of GN to build the library) as an example/starting point
for someone making a plugin outside of the FDE repository.

Addresses #114 for Linux.