WIP: Switch the C++ embedder code to a client wrapper
Rather than have a C++ interface in the shared library, make the C++
code a client wrapper layer that plugins and embedders consume.

Ideally in the future more code will move out of this layer and behind
a C interface in the library, so that the wrapper is as thin as
possible, but this addresses the immediate issue of the shared library
not having a viable interface.