Make the Windows example build self-contained
Copies the necessary Flutter data to a known location relative to the
executable, and uses executable-relative paths rather than
working-directory-relative paths.

This allows the executable to be run from any location, and makes it
clear what must be bundled as part of a working example. Follows the
same basic structure as the Linux example's bundling.