Add restat to Windows solink command
The link step includes the .lib file as output, but the link tool only
updates that file if something has changed, so it's possible to get into
situations where the .lib file is always older than the inputs, causing
the (expensive) link step to run every time.

This enabled ninja's 'restat' mode for the link step, so that it detects
the files that don't change and ignores them for the purposes of
determining what needs a rebuild.