[windows] Switch to GN for library build (#262)
Eliminates the VS-based embedder library build and updates the
example to use the GN build.

As part of the change, moves build output to
example/build/windows_fde
to match the recent change for Linux. Also eliminates the
'Dynamic Library' portion of the configuration names since
they are relics of a distinction that was removed in an earlier
change.