Update build to use FlutterMacOS.framework
What was FlutterEmbedderMac exists now only as a project to provide a
shared source for downloading the framework, and has been moved to
plugins/common/macos/SharedFlutterFramework.xcodeproj.

In order to avoid people thinking their apps should depend on that
project, the example instead fetches its own copy of the framework to a
different location. This is slightly wasteful, but simplifies
understanding of the example.