Move macOS plugin registration to a separate file (#330)
Matches the structure of the iOS GeneratedPluginRegistrant. Uses ObjC,
rather than Swift, so that changes to add a new plugin are localized to
one file, instead of also affecting the bridging header.

This will make it easier to eventually transition to generated plugin
registration, and in the meantime makes it slightly easier to add
plugins manually.