Support --track-widget-creation on Linux (#344)
Plumbs the flag for enabling widget creation tracking from the build
script through to 'flutter build bundle'. Fixes the crash-on-hot-reload
in VS Code for Linux.

Since this was touching the build.sh/Makefile interaction, also plumbs
through FLUTTER_ROOT, although doesn't yet rely on it being there in the
example.