Support --track-widget-creation on Windows
Plumbs the flag for enabling widget creation tracking from the build
script through to 'flutter build bundle'. Fixes the crash-on-hot-reload
in VS Code for Windows.

Since this added the logic to generate a property file for VS with
settings, also plumbs through FLUTTER_ROOT, although doesn't yet rely on
it being there in the example.