Add a min Flutter version to the example
This will help catch some errors caused by version mismatches that would
otherwise cause error messages that aren't at all self-explanatory.
This won't catch all cases, but for those that it does the failure will
be very early, and much clearer.

The intent is to rev this number every time a change is made on the
Flutter side that requires a corresponding change in this repository.