Add example_plugin implementation for macOS
For now this uses essentially the same structure as the other plugins,
although it doesn't use the shared project/framework copy, in order to
make it less closely tied to the FDE repo.

It still use the tools/ directory, but that will be eliminated as part
of migrating to CocoaPods in the future.