Fix some minor analyzer issues
Fixes a couple of issues that have slipped in, and removes a deprecated
setting.

Also removes the folder-local analysis_options.yaml files from the two
example directories, since they are intended at this point to be copied
and re-used elsewhere, so referring out to the root yaml doesn't make
sense.