Fix Windows project for paths with spaces
The arguments to the helper commands were being quoted, but not the
commands themselves, causing failures if the path to the project had
spaces in it.

Also re-adds the previously-dropped example/Runner.sln, to make the
workflow smoother when opening the project in Visual Studio (to avoid
being prompted to save the auto-synthesized solution on quit).

Part of the fix for https://github.com/flutter/flutter/issues/32792