Implement example_plugin for Windows
Adds an example_plugin implementation for Windows.

As with other platforms, the current build approach is a placeholder
which will be replaced with real dependency management in the future.
For now, the structure is that the plugin project is designed to only be
buildable from a Runner.sln; it gets its copy of the Flutter artifact
cache from the Runner's copy, and it builds into the Runner's build
directory.

As with the other platforms, this does not have dependencies on the rest
of FDE, should should be copyable as a starting point for making other
experimental plugins on Windows.

Fixes #357