Podify all macOS plugins
Converts all the macOS implementations of plugins to be pods. Removes
the projects and framework-related files, restructures them to follow
the Flutter plugin podspec convention, and adds podspec files.

The common definitions of the channel constants are removed in this
restructuring, since having source files that live outside the pod
directory is problematic. This means slightly more duplication, but the
constants should not be changing frequently (and this is a general
problem with Flutter plugins). The documentation that was in the
channel_constant.h files has moved to the appropirate Dart files.

Since the plugins are no longer including C++ constant definitions,
converted them to ObjC (.m) files.

Testbed now builds and loads all plugins via the Flutter tooling rather
than direct project depedencies.