Overhaul the primary READMEs
Streamline and simplify the READMEs:
- Moves the setup and `flutter run` documentation to the main page,
  condensing the setup description slightly.
- Moves the IDE setup to the main page; it's now much shorter, and
  slightly less step-by-step, but having it on the main page makes it
  much more discoverable.
- Adds a bit of discussion of how example/ is different from a stock
  template app to make the changes more discoverable.
- Eliminates Quick-Start, moving that content to the example README.

This makes the flow much simpler: the main README has all the
information to get started, including the command to actually run the
example. Then the example README has more information about how it works
and how it can be re-used.

Improving the plugin READMEs is left for a follow-up.