Convert macOS example_plugin to Swift
Uses the same structure as an iOS Flutter plugin using the Swift
template, with an ObjC interface wrapper.

This required adding use_frameworks to the Podfile, which will be added
to the template upstream as well.

Also adds the disable_input_output_paths that was added upstream to
address strict build issues with duplicate outputs.

Fixes #421