Convert plugins to Swift-based (#424)
Updates the example, testbed, and plugins on macOS to account for
https://github.com/flutter/flutter/pull/33997

example_plugin is now entirely Swift, with the ObjC wrapper removed; the
other plugins remain ObjC for now, with a Swift wrapper. The podspecs
have been adjusted to remove the public ObjC headers.

The example has been switched from an empty ObjC registrant to an empty
Swift registrant, and testbed has been updated with the results of
re-running `flutter run` with the new tooling support, and switched to
calling the Swift registrant as well.