Add a Warnings.xcconfig
Adds a new config file with a number of best-practice warnings. These
are in a config, rather than the project, to make it easier to maintain
over time in a template, as well as to see what was explicitly enabled
rather than default from the Xcode project.

Where the settings duplicate an existing setting in the project, it's
removed from the project so there's a single place that it's being set.

macOS portion of #343