Update Linux apps for incremental GLFW runloop (#496)
Switches to the new API from flutter/engine#11368 and moves the GTK event handling into the application level of 'testbed'. 'example' doesn't do that handling since it doesn't use plugins. (Longer term a specific toolkit like GTK will be a core requirement of the library, so this divergence won't exist; for now this makes the simple example easier to build and run.)

Note: Since the run-forever version in the C++ wrapper is still present (to avoid this change being a hard compilation break), and Windows didn't need special handling beyond what that does, Windows is not updated; it will be changing to a new API soon anyway, and leaving it avoids collisions with that work.