Commit rewrites from using current Xcode
Making any changes to the Xcode projects with 10.3 is causing these
diffs; they are presumably a combination of things that have changed in
how Xcode stores values over time, and things where we've hand-edited
the project in a way that's not exactly what Xcode would write.

Since the latent diffs causes annoyance for every change, this is a
no-op change to the projects that just allows Xcode to rewrite them the
way it wants to.