Move Windows window setup constants to a config
Rather than having the Window title and size be inlined in main.cpp,
where local modifications will cause conflicts on every API update,
move them to a separate constants file. This will help isolate early
users of the Windows embedding from API cuhrn as we iterate on the
Windows Flutter API surface.

Per the comment in the file this is not a long-term solution to
configuration, but it should help in the short term.