Use solution-level Flutter cache for Windows plugin builds (#561)
Incrementally improves the Windows build structure by having the plugins use the flutter\ephemeral\ cache from the application project, rather than doing their own unpack copies. This requires moving the backend_tool call to a separate project, rather than a pre-build step, so that plugin projects can be made to depend on it in the solution to enforce ordering.