Share ephemeral dir for Linux plugin builds
Rather than having each plugin 'unpack' a copy of the Flutter artifacts,
make the application build responsible for populating that directory
before running plugin builds, and passing that directory to the plugin
Makefiles.

This eliminates the undesirable direct use of 'unpack' on Linux, and
makes the structure of the build more consistent with Windows.

Also opportunitistically fixes an issue caught while changing
neighboring lines where OBJ_DIR was not including the build type, which
would cause unexpected sharing of output files with different flags.