[windows] Add SDK to FlutterBuild
Even though the SDK isn't used by this project, VS still appears to do
SDK validation when the build starts (as with the tool version).

Longer term, the ideal solution would be to find a project template that
doesn't require setting tool and SDK versions that we don't need, if
possible, since all it needs to do is run a script.