Covert file_chooser API to Futures
file_chooser was written with an ObjC-like callback API, when an API
based on futures would be more idiomatic for Dart. This converts to a
Futures-based API.

Also changes the implementation so that missing plugin exceptions will
be propagated to clients for handling, rather than just logged.

This is a breaking change for clients of the file_chooser plugin.

Fixes #617