Use plugin symlinks on Windows and Linux (#670)
Switches Windows and Linux to using the new plugin symlinks. This eliminates
the need for project files/makefiles to have developer-specific paths
(especially if pulled via pub).

Now that the plugin plumbing is mostly generic for Linux, most of it in moved
into the example Makefile so that the changes necessary to use plugins are
much more targeted. Updates the documentation for using plugins on Linux
to be more specific now that the process is more straightforward.

Part of #578