Use generated plugin makefile on Linux (#686)
The Flutter tool now generates a makefile suitable for inclusion in the
main makefile that adds all the necessary plugin flags and targets.
This switches to using that new system.

The header copy step is removed from the plugins, since the
generated makefile adds the necessary include paths.

Linux plugins no longer require manual build integration, so
those instructions are removed. Since there's now no need
for instructions in most cases, this removes most of the plugin
instructions. The sections about building plugins manually
are removed since there's no evidence that people are doing
manual add-to-app experimentation at this point, and it's
not something that we're supporting enough to encourage
it with partial instructions. The result is a much simpler set
of plugin instructions, which should be more approachable.