Adjust plugin metada
Plugins from FDE should not be published. Add a hurdle so that anyone
who thinks doing so is a good idea, against the explicit guidance of
both this project and the Flutter for dekstop documation, has to go out
of their way to do so.

Also removes the deprecated author, and the homepage (since no published
plugin should ever be linking here).