Add preliminary CMake build for GTK (#735)
Adds a build for the in-progress GTK runner using the new CMake tooling,
as well as the new unpacking of GTK artifacts. Unlike the previous
Make-based implementation, this is stand-alone, rather than requiring
bootstrapping with the GLFW build first. Also doesn't require using a
local engine build.

Plugins are not yet supported, but TODOs are left for wiring them in.